requirements.txt
aiohttp==2.0.7
aiohttp-jinja2==0.13.0
injector==0.12.1
pybase62==0.2.0
asyncpg==0.10.1
pytest==3.0.7
